refactor: introduce macro for standard filter builder methods across request types
